+++ tests/e2e/utils/src/components.js
@@ -7,10 +7,8 @@
  */
 import { merchant, IS_RETEST_MODE } from './flows';
 import {
-	clickTab,
 	uiUnblocked,
 	verifyCheckboxIsUnset,
-	selectOptionInSelect2,
 	setCheckbox,
 	unsetCheckbox,
 	evalAndClick,
@@ -24,6 +22,8 @@ const client = factories.api.withDefaultPermalinks;
 const config = require( 'config' );
 const simpleProductName = config.get( 'products.simple.name' );
 const simpleProductPrice = config.has('products.simple.price') ? config.get('products.simple.price') : '9.99';
+const defaultVariableProduct = config.get('products.variable');
+const defaultGroupedProduct = config.get('products.grouped');
 
 /**
  * Verify and publish
@@ -135,6 +135,9 @@ const completeOnboardingWizard = async () => {
 
 	// Business Details section
 
+	// Temporarily add delay to reduce test flakiness
+	await page.waitFor( 2000 );
+
 	// Query for the <SelectControl>s
 	const selectControls = await page.$$( '.woocommerce-select-control' );
 	expect( selectControls ).toHaveLength( 2 );
@@ -220,172 +223,104 @@ const createSimpleProductWithCategory = async ( productName, productPrice, categ
 
 /**
  * Create variable product.
+ * Also, create variations for all attributes.
+ *
+ * @param varProduct Defaults to the variable product object in `default.json`
+ * @returns the ID of the created variable product
  */
-const createVariableProduct = async () => {
+const createVariableProduct = async (varProduct = defaultVariableProduct) => {
+	const { attributes } = varProduct;
+	const { id } = await factories.products.variable.create(varProduct); // create the variable product
+	const variations = [];
+	const buffer = []; // accumulated attributes while looping
+	const aIdx = 0; // attributes[] index
 
-	// We need to remove any listeners on the `dialog` event otherwise we can't catch the dialogs below
-	page.removeAllListeners('dialog');
+	// Create variation for all attributes
+	const createVariation = (aIdx) => {
+		const { name, options } = attributes[aIdx];
+		const isLastAttribute = aIdx === attributes.length - 1;
 
-	// Go to "add product" page
-	await merchant.openNewProduct();
+		// Add each attribute value to the buffer.
+		options.forEach((opt) => {
+			buffer.push({
+				name: name,
+				option: opt
+			});
 
-	// Make sure we're on the add product page
-	await expect( page.title() ).resolves.toMatch( 'Add new product' );
+			if (isLastAttribute) {
+				// If this is the last attribute, it means the variation is now complete.
+				// Save whatever's been accumulated in the buffer to the `variations[]` array.
+				variations.push({
+					attributes: [...buffer]
+				});
+			} else {
+				// Otherwise, move to the next attribute first
+				// before proceeding to the next value in this attribute.
+				createVariation(aIdx + 1);
+			}
 
-	// Set product data
-	await expect( page ).toFill( '#title', 'Variable Product with Three Variations' );
-	await expect( page ).toSelect( '#product-type', 'Variable product' );
+			buffer.pop();
+		});
+	};
+	createVariation(aIdx);
 
-	// Create attributes for variations
-	await clickTab( 'Attributes' );
-	await expect( page ).toSelect( 'select[name="attribute_taxonomy"]', 'Custom product attribute' );
+	// Set some properties of 1st variation
+	variations[0].regularPrice = '9.99';
+	variations[0].virtual = true;
 
-	for ( let i = 0; i < 3; i++ ) {
-		await expect( page ).toClick( 'button.add_attribute', { text: 'Add' } );
-		// Wait for attribute form to load
-		await uiUnblocked();
+	// Set some properties of 2nd variation
+	variations[1].regularPrice = '11.99';
+	variations[1].virtual = true;
 
-		await page.focus( `input[name="attribute_names[${ i }]"]` );
-		await expect( page ).toFill( `input[name="attribute_names[${ i }]"]`, 'attr #' + ( i + 1 ) );
-		await expect( page ).toFill( `textarea[name="attribute_values[${ i }]"]`, 'val1 | val2' );
-		await expect( page ).toClick( `input[name="attribute_variation[${ i }]"]` );
+	// Set some properties of 3rd variation
+	variations[2].regularPrice = '20';
+	variations[2].weight = '200';
+	variations[2].dimensions = {
+		length: '10',
+		width: '20',
+		height: '15'
+	};
+	variations[2].manage_stock = true;
+
+	// Use API to create each variation
+	for (const v of variations) {
+		await factories.products.variation.create({
+			productId: id,
+			variation: v
+		});
 	}
 
-	await expect( page ).toClick( 'button', { text: 'Save attributes' } );
-
-	// Wait for attribute form to save (triggers 2 UI blocks)
-	await uiUnblocked();
-	await page.waitFor( 1000 );
-	await uiUnblocked();
-
-	// Create variations from attributes
-	await clickTab( 'Variations' );
-	await page.waitForSelector( 'select.variation_actions:not([disabled])' );
-	await page.focus( 'select.variation_actions' );
-	await expect( page ).toSelect( 'select.variation_actions', 'Create variations from all attributes' );
-
-	const firstDialog = await expect( page ).toDisplayDialog( async () => {
-		// Using this technique since toClick() isn't working.
-		// See: https://github.com/GoogleChrome/puppeteer/issues/1805#issuecomment-464802876
-		page.$eval( 'a.do_variation_action', elem => elem.click() );
-
-	} );
-
-	expect( firstDialog.message() ).toMatch( 'Are you sure you want to link all variations?' );
-
-	const secondDialog = await expect( page ).toDisplayDialog( async () => {
-		await firstDialog.accept();
-	} );
-
-	expect( secondDialog.message() ).toMatch( '8 variations added' );
-	await secondDialog.dismiss();
-
-	// Set some variation data
-	await uiUnblocked();
-	await uiUnblocked();
-
-	await page.waitForSelector( '.woocommerce_variation .handlediv' );
-
-	// Verify that variations were created
-	await Promise.all( [
-		expect( page ).toMatchElement( 'select[name="attribute_attr-1[0]"]', { text: 'val1' } ),
-		expect( page ).toMatchElement( 'select[name="attribute_attr-2[0]"]', { text: 'val1' } ),
-		expect( page ).toMatchElement( 'select[name="attribute_attr-3[0]"]', { text: 'val1' } ),
-
-		expect( page ).toMatchElement( 'select[name="attribute_attr-1[1]"]', { text: 'val1' } ),
-		expect( page ).toMatchElement( 'select[name="attribute_attr-2[1]"]', { text: 'val1' } ),
-		expect( page ).toMatchElement( 'select[name="attribute_attr-3[1]"]', { text: 'val2' } ),
-
-		expect( page ).toMatchElement( 'select[name="attribute_attr-1[2]"]', { text: 'val1' } ),
-		expect( page ).toMatchElement( 'select[name="attribute_attr-2[2]"]', { text: 'val2' } ),
-		expect( page ).toMatchElement( 'select[name="attribute_attr-3[2]"]', { text: 'val1' } ),
-
-		expect( page ).toMatchElement( 'select[name="attribute_attr-1[3]"]', { text: 'val1' } ),
-		expect( page ).toMatchElement( 'select[name="attribute_attr-2[3]"]', { text: 'val2' } ),
-		expect( page ).toMatchElement( 'select[name="attribute_attr-3[3]"]', { text: 'val2' } ),
-
-		expect( page ).toMatchElement( 'select[name="attribute_attr-1[4]"]', { text: 'val2' } ),
-		expect( page ).toMatchElement( 'select[name="attribute_attr-2[4]"]', { text: 'val1' } ),
-		expect( page ).toMatchElement( 'select[name="attribute_attr-3[4]"]', { text: 'val1' } ),
-
-		expect( page ).toMatchElement( 'select[name="attribute_attr-1[5]"]', { text: 'val2' } ),
-		expect( page ).toMatchElement( 'select[name="attribute_attr-2[5]"]', { text: 'val1' } ),
-		expect( page ).toMatchElement( 'select[name="attribute_attr-3[5]"]', { text: 'val2' } ),
-
-		expect( page ).toMatchElement( 'select[name="attribute_attr-1[6]"]', { text: 'val2' } ),
-		expect( page ).toMatchElement( 'select[name="attribute_attr-2[6]"]', { text: 'val2' } ),
-		expect( page ).toMatchElement( 'select[name="attribute_attr-3[6]"]', { text: 'val1' } ),
-
-		expect( page ).toMatchElement( 'select[name="attribute_attr-1[7]"]', { text: 'val2' } ),
-		expect( page ).toMatchElement( 'select[name="attribute_attr-2[7]"]', { text: 'val2' } ),
-		expect( page ).toMatchElement( 'select[name="attribute_attr-3[7]"]', { text: 'val2' } ),
-	] );
-
-	await expect( page ).toClick( '.woocommerce_variation:nth-of-type(2) .handlediv' );
-	await page.waitFor( 2000 );
-	await page.focus( 'input[name="variable_is_virtual[0]"]' );
-	await expect( page ).toClick( 'input[name="variable_is_virtual[0]"]' );
-	await expect( page ).toFill( 'input[name="variable_regular_price[0]"]', '9.99' );
-
-	await expect( page ).toClick( '.woocommerce_variation:nth-of-type(3) .handlediv' );
-	await page.waitFor( 2000 );
-	await page.focus( 'input[name="variable_is_virtual[1]"]' );
-	await expect( page ).toClick( 'input[name="variable_is_virtual[1]"]' );
-	await expect( page ).toFill( 'input[name="variable_regular_price[1]"]', '11.99' );
-
-	await expect( page ).toClick( '.woocommerce_variation:nth-of-type(4) .handlediv' );
-	await page.waitFor( 2000 );
-	await page.focus( 'input[name="variable_manage_stock[2]"]' );
-	await expect( page ).toClick( 'input[name="variable_manage_stock[2]"]' );
-	await expect( page ).toFill( 'input[name="variable_regular_price[2]"]', '20' );
-	await expect( page ).toFill( 'input[name="variable_weight[2]"]', '200' );
-	await expect( page ).toFill( 'input[name="variable_length[2]"]', '10' );
-	await expect( page ).toFill( 'input[name="variable_width[2]"]', '20' );
-	await expect( page ).toFill( 'input[name="variable_height[2]"]', '15' );
-
-	await page.focus( 'button.save-variation-changes' );
-	await expect( page ).toClick( 'button.save-variation-changes', { text: 'Save changes' } );
-
-	await verifyAndPublish( 'Product published.' );
-
-	const variablePostId = await page.$( '#post_ID' );
-	let variablePostIdValue = ( await ( await variablePostId.getProperty( 'value' ) ).jsonValue() );
-	return variablePostIdValue;
+	return id;
 };
 
 /**
  * Create grouped product.
+ * 
+ * @param groupedProduct Defaults to the grouped product object in `default.json`
+ * @returns ID of the grouped product
  */
-const createGroupedProduct = async () => {
-	// Create two products to be linked in a grouped product after
-	await factories.products.simple.create( {
-		name: simpleProductName + ' 1',
-		regularPrice: simpleProductPrice
-	} );
-	await factories.products.simple.create( {
-		name: simpleProductName + ' 2',
-		regularPrice: simpleProductPrice
-	} );
+const createGroupedProduct = async (groupedProduct = defaultGroupedProduct) => {
+	const { name, groupedProducts } = groupedProduct;
+	const simpleProductIds = [];
+	let groupedProductRequest;
 
-	// Go to "add product" page
-	await merchant.openNewProduct();
+	// Using the api, create simple products to be grouped
+	for (const simpleProduct of groupedProducts) {
+		const { id } = await factories.products.simple.create(simpleProduct);
+		simpleProductIds.push(id);
+	}
 
-	// Make sure we're on the add product page
-	await expect( page.title() ).resolves.toMatch( 'Add new product' );
+	// Using the api, create the grouped product
+	groupedProductRequest = {
+		name: name,
+		groupedProducts: simpleProductIds
+	};
+	const { id } = await factories.products.grouped.create(
+		groupedProductRequest
+	);
 
-	// Set product data and save the product
-	await expect( page ).toFill( '#title', 'Grouped Product' );
-	await expect( page ).toSelect( '#product-type', 'Grouped product' );
-	await clickTab( 'Linked Products' );
-	await selectOptionInSelect2( simpleProductName + ' 1' );
-	await selectOptionInSelect2( simpleProductName + ' 2' );
-	await verifyAndPublish();
-
-	// Get product ID
-	const groupedPostId = await page.$( '#post_ID' );
-	let groupedPostIdValue = ( await ( await groupedPostId.getProperty( 'value' ) ).jsonValue() );
-	return groupedPostIdValue;
-}
+	return id;
+};
 
 /**
  * Create a basic order with the provided order status.